+++ CHANGES.md
@@ -3,6 +3,7 @@
 * dependsOn can now refer to any field, not just another dropdown. So you can create a dropdown depending on a text field too (as an example)
 * Added trigger "toolbarsearch", so you can hook into this if need be
 * Added general option resetTable, can be used if sortingInfoSelector is used, to set an extra button to reset the table to its original state (paging, sorting, visibility)
+* flatpickr also supported as datepicker (also returns Y-m-d to the backend)
 
 = 3.1.3 (2025/07/15) =
 * the toolbarsearch extension is now in the main plugin